remove unnecessary comments, fix bash shell is executed again when we restart the container

+++ shells/init-openvpn.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
+
 mkdir -p -v /dev/net
 mknod /dev/net/tun c 10 200
 chmod 600 /dev/net/tun
 echo "device created"
 
-bash serverConfig.sh
-bash baseClientConfig.sh
-bash generateClientConfig.sh
-
-ln -s `pwd`/generateClientConfig.sh /usr/bin/
+if [ ! -f ${EASYRSA_PKI}/ca.crt ]
+then 
+    bash serverConfig.sh
+    bash baseClientConfig.sh
+    bash generateClientConfig.sh    
+    ln -s `pwd`/generateClientConfig.sh /usr/bin/
+fi
 
 openvpn --config ${OVPN_DIR}/server.conf
 
-#this line prevent docker to stop immediately
 exec "$@"
-#while getopts 
+
+#commit message "remove unnecessary comments, fix bash shell is executed again when we restart the container"
